+++ src/oscar/Graphics/RenderBuffer.h
@@ -1,16 +1,17 @@
 #pragma once
 
 #include <oscar/Graphics/RenderBufferType.h>
-#include <oscar/Graphics/RenderTextureDescriptor.h>
 
 #include <memory>
+
+namespace osc { struct RenderTextureParams; }
 
 namespace osc
 {
     class RenderBuffer final {
     public:
         RenderBuffer(
-            const RenderTextureDescriptor&,
+            const RenderTextureParams&,
             RenderBufferType
         );
         RenderBuffer(const RenderBuffer&);