+++ src/oscar_learnopengl/PBR/LOGLPBRSpecularIrradianceTab.cpp
@@ -54,9 +54,11 @@ namespace
         hdr_texture.set_wrap_mode(TextureWrapMode::Clamp);
         hdr_texture.set_filter_mode(TextureFilterMode::Linear);
 
-        RenderTexture cubemap_render_target{{512, 512}};
-        cubemap_render_target.set_dimensionality(TextureDimensionality::Cube);
-        cubemap_render_target.set_color_format(RenderTextureFormat::RGBFloat16);
+        RenderTexture cubemap_render_target{{
+            .dimensions = {512, 512},
+            .dimensionality = TextureDimensionality::Cube,
+            .color_format = RenderTextureFormat::RGBFloat16,
+        }};
 
         // create a 90 degree cube cone projection matrix
         const Mat4 projection_matrix = perspective(90_deg, 1.0f, 0.1f, 10.0f);
@@ -85,9 +87,11 @@ namespace
         IResourceLoader& loader,
         const RenderTexture& skybox)
     {
-        RenderTexture irradiance_cubemap{{32, 32}};
-        irradiance_cubemap.set_dimensionality(TextureDimensionality::Cube);
-        irradiance_cubemap.set_color_format(RenderTextureFormat::RGBFloat16);
+        RenderTexture irradiance_cubemap{{
+            .dimensions = {32, 32},
+            .dimensionality = TextureDimensionality::Cube,
+            .color_format = RenderTextureFormat::RGBFloat16,
+        }};
 
         const Mat4 capture_projection = perspective(90_deg, 1.0f, 0.1f, 10.0f);
 
@@ -117,9 +121,11 @@ namespace
         const int level_zero_width = 128;
         static_assert(std::popcount(static_cast<unsigned>(level_zero_width)) == 1);
 
-        RenderTexture capture_render_texture{{level_zero_width, level_zero_width}};
-        capture_render_texture.set_dimensionality(TextureDimensionality::Cube);
-        capture_render_texture.set_color_format(RenderTextureFormat::RGBFloat16);
+        RenderTexture capture_render_texture{{
+            .dimensions = {level_zero_width, level_zero_width},
+            .dimensionality = TextureDimensionality::Cube,
+            .color_format = RenderTextureFormat::RGBFloat16,
+        }};
 
         const Mat4 capture_projection = perspective(90_deg, 1.0f, 0.1f, 10.0f);
 
@@ -163,8 +169,10 @@ namespace
     Texture2D create_2D_brdf_lookup(
         IResourceLoader& loader)
     {
-        RenderTexture render_texture{{512, 512}};
-        render_texture.set_color_format(RenderTextureFormat::RGFloat16);
+        RenderTexture render_texture{{
+            .dimensions = {512, 512},
+            .color_format = RenderTextureFormat::RGFloat16,
+        }};
 
         Material material{Shader{
             loader.slurp("oscar_learnopengl/shaders/PBR/ibl_specular/BRDF.vert"),
@@ -321,7 +329,7 @@ private:
     RenderTexture irradiance_map_ = create_irradiance_cubemap(loader_, projected_map_);
     Cubemap prefilter_map_ = create_prefiltered_environment_map(loader_, projected_map_);
     Texture2D brdf_lookup_ = create_2D_brdf_lookup(loader_);
-    RenderTexture output_render_texture_{{1, 1}};
+    RenderTexture output_render_texture_;
 
     Material background_material_{Shader{
         loader_.slurp("oscar_learnopengl/shaders/PBR/ibl_specular/Skybox.vert"),