+++ CHANGELOG.md
@@ -8,9 +8,10 @@ on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added `Point Forces` and `Point Torques` visualization options, which enables drawing force
   vectors in their point-force form (as opposed to their reduced body-force form). This feature
   currently only works for `ExternalForce`s in the model (#904) and `GeometryPath`s (#907).
-
 - The `Show Forces' Linear/Rotation Component` option was reworded to `Forces on Bodies` and
   `Torques on Bodies` to reflect what's actually being shown
+- Internal: `RenderTextureDescriptor` was refactored into a POD struct called `RenderTextureParams`
+  for ease-of-use
 
 ## [0.5.13] - 2024/07/30